+++ src/bindings/value.rs
@@ -1,4 +1,4 @@
-use libquickjs_sys as q;
+use libquickjspp_sys as q;
 
 use crate::{ExecutionError, JsValue, ValueError};